Moved tests out of src/ and into test/.
Additionally, the test classes now belong to the same packages
as the classes they are testing.

The sole exception to this is the Beauti test class, which requires
access to package private methods/fields from its parent class,
which belongs to the beast core.  Once beast core tests migrate to
this structure (being optimistic here) this won't be necessary.